Allow multiple users
Instead of regenerating the htpasswd file, it now appends to it to allow for multiple users to be added.

This might need more intelligent updating of existing users instead of just appending them. Also one could argue that there should be a way to remove a user from the auth. (including a warning if auth=on but there are no users left to match against)

[ ] update existing users
[ ] allow for removing users
[ ] if no users remain in the passwd file, ask what to do (remove auth, leave locked with no way of getting in or immediately ask for a new username/password