refactor: use path available to nginx process for htpasswd files
Nginx worker processes do not have access to the dokku user's home directory, so they need to be written to a path under /etc/nginx. The nginx conf.d files _will_ be loaded as they are read by the root process (except for SELinux), so while the config gets loaded, the actual htpasswd files will not.

For management purposes, we move the actual files to a new data directory, and then copy it out from there to the /etc/nginx/htpasswd directory for usage by nginx.

Closes #24